+++ python_modules/dagster/dagster_tests/components_tests/executable_component_tests/test_no_param_assets.py
@@ -1,82 +1,29 @@
 from collections import defaultdict
-from collections.abc import Iterator
-from contextlib import contextmanager
-from typing import Any, Optional
 
 import pytest
-from dagster._core.definitions.asset_key import AssetKey, CoercibleToAssetKey
+from dagster._core.definitions.asset_key import AssetKey
 from dagster._core.definitions.asset_spec import AssetSpec
-from dagster._core.definitions.definitions_class import Definitions
-from dagster._core.definitions.events import AssetMaterialization
 from dagster._core.definitions.resource_annotation import ResourceParam
 from dagster._core.definitions.result import MaterializeResult
-from dagster._core.events import DagsterEvent
 from dagster._core.execution.context.asset_execution_context import AssetExecutionContext
 from dagster._core.execution.context.input import InputContext
 from dagster._core.execution.context.output import OutputContext
 from dagster._core.storage.io_manager import IOManager
-from dagster.components.core.context import ComponentLoadContext
-from dagster.components.lib.executable_component import ExecutableComponent, make_materialize_result
+from dagster.components.lib.executable_component.executable_component import (
+    ExecutableComponent,
+    make_materialize_result,
+)
+from dagster.components.lib.executable_component.execute import (
+    components_to_defs,
+    execute_asset_in_component,
+    execute_assets_in_component,
+    execute_assets_in_defs,
+)
+from dagster.components.lib.executable_component.load_input import load_input
 
 
 def test_single_asset():
     assert ExecutableComponent
-
-
-class MaterializationEvaluation:
-    def __init__(self, event: AssetMaterialization, value: Any):
-        self.event = event
-        self.value = value
-
-
-def execute_asset_in_component(
-    component: ExecutableComponent,
-    asset_key: CoercibleToAssetKey,
-    resources: Optional[dict[str, Any]] = None,
-) -> MaterializationEvaluation:
-    asset_key = AssetKey.from_coercible(asset_key)
-    return execute_assets_in_component(component, [asset_key], resources)[asset_key]
-
-
-def execute_assets_in_component(
-    component: ExecutableComponent,
-    asset_keys: list[CoercibleToAssetKey],
-    resources: Optional[dict[str, Any]] = None,
-) -> dict[AssetKey, MaterializationEvaluation]:
-    return execute_assets_in_defs(
-        component.build_defs(ComponentLoadContext.for_test()), asset_keys, resources
-    )
-
-
-def execute_assets_in_defs(
-    defs: Definitions,
-    asset_keys: list[CoercibleToAssetKey],
-    resources: Optional[dict[str, Any]] = None,
-) -> dict[AssetKey, MaterializationEvaluation]:
-    full_asset_keys: list[AssetKey] = [AssetKey.from_coercible(key) for key in asset_keys]
-    if resources:
-        defs = defs.with_resources(resources)
-    job_def = defs.get_implicit_job_def_for_assets(full_asset_keys)
-    assert job_def
-
-    result = job_def.execute_in_process()
-    assert result.success
-
-    evaluations = {}
-    for asset_key in full_asset_keys:
-        node_output_handle = job_def.asset_layer.node_output_handle_for_asset(asset_key)
-        node_name = node_output_handle.node_handle.name
-
-        materializations = result.asset_materializations_for_node(node_name)
-        for materialization in materializations:
-            evaluations[asset_key] = MaterializationEvaluation(
-                event=materialization,
-                value=result.output_for_node(
-                    node_name, output_name=asset_key.to_python_identifier()
-                ),
-            )
-
-    return evaluations
 
 
 def test_basic_single_asset_return_none() -> None:
@@ -137,12 +84,6 @@ def test_assets_with_resources() -> None:
     assert evaluation.value == "foo"
 
 
-def components_to_defs(components: list[ExecutableComponent]) -> Definitions:
-    return Definitions.merge(
-        *[component.build_defs(ComponentLoadContext.for_test()) for component in components]
-    )
-
-
 def test_deps() -> None:
     def _upstream_execute(context) -> MaterializeResult:
         return make_materialize_result()
@@ -165,23 +106,6 @@ def test_deps() -> None:
     defs = components_to_defs([upstream_component, downstream_component])
     evaluations = execute_assets_in_defs(defs, ["upstream_key", "downstream_key"])
     assert set(evaluations.keys()) == {AssetKey("upstream_key"), AssetKey("downstream_key")}
-
-
-def resource_dict(context: AssetExecutionContext) -> dict[str, Any]:
-    return context.op_execution_context.resources.original_resource_dict
-
-
-@contextmanager
-def load_input(context: AssetExecutionContext, input_name: str) -> Iterator[Any]:
-    step_context = context.get_step_execution_context()
-    step_input = step_context.step.step_input_named(input_name)
-    input_def = step_context.op_def.input_def_named(input_name)
-    iterator = step_input.source.load_input_object(step_context, input_def)
-    for item in iterator:
-        if isinstance(item, DagsterEvent):
-            context.op_execution_context._events.append(item)  # noqa: SLF001
-        else:
-            yield item
 
 
 def test_parameter_passing_using_user_space_io_manager() -> None:
@@ -207,7 +131,7 @@ def test_parameter_passing_using_user_space_io_manager() -> None:
     def _downstream_execute(
         context: AssetExecutionContext,
     ) -> MaterializeResult:
-        with load_input(context, "upstream_key") as upstream_key_value:
+        with load_input(context, "upstream_key", int) as upstream_key_value:
             return make_materialize_result(value=upstream_key_value + 1)
 
     upstream_component = ExecutableComponent(
